refactor(alwatr-on): rename lib
